==> InstaCart_Project/Notebook Files/06_Using Test Data.ipynb <==
import pandas as pd
import json 
import numpy as np
import pathlib

# Read in test data from json format

test_data_dir = pathlib.Path('test_data/')
test_data_files = list(test_data_dir.glob('*.json'))

all_test_data = {}
for test_data_file in test_data_files: 
    with open(test_data_file) as fin:
        all_test_data[test_data_file.name.replace('.json', '')] = json.load(fin)
        
all_test_data['test_1'].keys()

## Convert To Dataframe

table_values = []
user = 12345678912345
order_id = 98765432198765
order_num = 0
for test in all_test_data:
    test_order_history = all_test_data[test]['order_history']
    for order in test_order_history:
        order_num += 1 
        product_list = test_order_history[order]['products']
        dow = test_order_history[order]['order_dow']
        hour= test_order_history[order]['order_hour_of_day']
        days = test_order_history[order]['days_since_prior_order']
        for product in product_list:
            table_values.append([user, order_id, order, product, dow, hour, days])
        order_id+=1
    user +=1
   
test_data = pd.DataFrame(table_values, columns=['user_id', 'order_id', 'order', 'product_id', 'order_dow', 'order_hour_of_day', 'days_since_prior_order'])

test_data.shape

products_csv = pd.read_csv('products.csv')

test_data = pd.merge(test_data, products_csv, how = 'inner', on = 'product_id')

test_data.columns

test_data.drop(['product_id', 'aisle_id', 'department_id'], axis = 1, inplace = True)

test_data.shape

### Create User mean profile for clustering purposes 

test_user_mean = test_data.groupby('user_id')['order_dow','order_hour_of_day', 'days_since_prior_order'].agg(np.nanmean)
total_orders = test_data.groupby('user_id').order_id.nunique() # <- series, will need to be added to larger dataframe 
# Average number of products in each order 
num_products = test_data.groupby('user_id').product_name.nunique()

# add average number of products in each order to user_mean df
test_user_mean['num_orders'] = total_orders
test_user_mean['num_products'] = num_products
test_user_mean['avg_num_products_per_order'] = num_products / total_orders

# Add every product ever ordered for each user 
# Combine users products into a list specific to that user   
product_list = []
for product in test_data.groupby('user_id')['product_name']:
    product_list.append(' '.join(product[1]))

test_user_mean['Products'] = product_list   
    

# 
test_user_mean.to_csv('test_user_profile.csv')

# test_data.to_csv('project_order_history_test_data.csv')

### Get the current order products into a dataframe with user id and order id and list of products in one row

current_table_values = []
user = 12345678912345
order_id = 98765432198765
for test in all_test_data:
    test_current_order = all_test_data[test]['current_order']
    for order in test_current_order:
        product_list = test_current_order[order]['products']
        dow = test_current_order[order]['order_dow']
        hour= test_current_order[order]['order_hour_of_day']
        days = test_current_order[order]['days_since_prior_order']
        for product in product_list:
            current_table_values.append([user, order_id, order, product, dow, hour, days])
    user +=1
    order_id+=1
current_test_data = pd.DataFrame(current_table_values, columns=['user_id', 'order_id', 'order', 'product_id', 'order_dow', 'order_hour_of_day', 'days_since_prior_order'])

current_test_data = pd.merge(current_test_data, products_csv, how = 'inner', on = 'product_id')

current_test_data.drop(['product_id', 'aisle_id', 'department_id'], axis = 1, inplace = True)

## Get a list of the products in one row of the dataframe

products = current_test_data['product_name']

product_no_space = []
for product in products:
    product = product.replace(" ", "_")
    product_no_space.append(product)

# drop original column, replace it with one with no space
current_test_data.drop(['product_name'], axis=1)
current_test_data['product_name'] = product_no_space

# add product name to each user
name_list = []
for p_name in current_test_data.groupby('order_id')['product_name']:
    name_list.append(' '.join(p_name[1]))
    
order_id = current_test_data.groupby('order_id')['product_name'].agg('count').index
# user_id = current_test_data.groupby(['user_id'])['product_name'].agg('count').index
user_id = current_test_data['user_id'].drop_duplicates()
order_id = current_test_data['order_id'].drop_duplicates()
current_test_order_products = pd.DataFrame({'user_id': user_id, 'order_id':order_id, 'products':name_list})


current_test_order_products.head()

###### Will need to get the xy coordinates of each new test user. calculate the euclidiean distance and use that to assign to each cluster

### Create a dictionary of current order information 

current_order_products_list = []
for row in current_test_order_products['products']:
    productsName = row.split(' ')
    current_order_products_list.append(productsName)
#     current_test_order_products['products']

current_test_order_products.head()

current_test_order_products['product_list'] = current_order_products_list

userid_cluster = pd.DataFrame(current_test_order_products['user_id'])

cluster_list = ['Cluster_6','Cluster_8','Cluster_5','Cluster_8','Cluster_1','Cluster_6','Cluster_8','Cluster_11','Cluster_10',
                'Cluster_12','Cluster_6','Cluster_12','Cluster_5','Cluster_4','Cluster_6','Cluster_1','Cluster_6','Cluster_8',
                'Cluster_1','Cluster_5','Cluster_13','Cluster_6','Cluster_4','Cluster_6','Cluster_14']

current_test_order_products['Clusters'] = cluster_list

current_test_order_products

current_order_dict= {} # products, product_list 
for index, content, in current_test_order_products.iterrows():
    user_id = content[0]
    order_id = content[1]
    product_list = content[3]
    cluster = content[4]
    if len(product_list) > 1:
        product_one = product_list[0]
        product_two = product_list[1]
        current_order_dict[user_id] = {'order_id': order_id, 'product_one': product_one, 'product_two': product_two, 'Cluster': cluster}
    else:
        product_one = product_list[0]
        current_order_dict[user_id] = {'order_id': order_id, 'product_one': product_one, 'Cluster': cluster}
    

with open('current_order_test_data.json', 'w') as file:
    json.dump(current_order_dict, file)
