==> InstaCart_Project/Notebook Files/08_Assign Test Data to Clusters.ipynb <==
# Classifies (Assign) Testing Data into our Clusters

import pandas as pd
from User_Clustering import kmeans  # brings in our kmeans clustering

### PCA on new testing data

user_profile = pd.read_csv('complete_user_profile_test.csv')
user_profile.head()
# Remove user_id to allow for PCA 
user_profile_no_id = user_profile.loc[:, 'order_dow':]

from sklearn.decomposition import PCA 

# perform PCA using 2 components 
pca = PCA(n_components = 2).fit(user_profile_no_id)
pca_2d = pca.transform(user_profile_no_id)
pca_df = pd.DataFrame(pca_2d)

predictions = kmeans.predict(pca_df)

user_profile['Cluster'] = predictions

user_profile
