==> InstaCart_Project/Notebook Files/09_Recommendation Main File.ipynb <==
import json
from Recommendation_Logic_Functions import getItemFromAllUsers, getItemFromCluster, getCurrentItems, getDefaultRecommendation
import pprint

current_test_data = {}

with open('current_order_test_data.json') as fin:
    current_test_data = json.load(fin)
    
user_bigram_dict = {}

with open('all_bigrams.json') as fin:
    user_bigram_dict = json.load(fin)
    
cluster_bigram_dict = {}

with open('all_cluster_bigrams.json') as fin:
    cluster_bigram_dict = json.load(fin)

# Create Main file variables
### Use for loop

for user, value in current_test_data.items():
#     print(user,"____", value)
    user_id = str(user)
    order_id = value['order_id']
#     print(user_id,"----", order_id)
    cart_list = getCurrentItems(user_id, order_id, current_test_data)
#     print(cart_list)
    user_cluster = value['Cluster']
#     print(user_cluster)
    
    rec = getItemFromCluster(cluster_bigram_dict, cart_list, user_cluster)
    if rec != 'None':
        print("User ",user_id, "'s Cart", cart_list, " Recommendation -------> ", rec, "  Source: Cluster Assignment")
    
    if rec == 'None':
        rec = getItemFromAllUsers(user_bigram_dict, cart_list)
        if rec != 'None':
            print("User ",user_id, "'s Cart ", cart_list, " Recommendation -------> ", rec, "  Source: All Users")
        
    if rec == 'None':
        rec = getDefaultRecommendation(cluster_bigram_dict, cart_list, user_cluster)
        if rec != 'None':
            print("User ",user_id, "'s Cart ", cart_list, " Recommendation -------> ", rec, "  Source: Default")
    if rec == 'None':
        print('*****************RECOMMENDATION SYSTEM DIDN\'T WORK********************')
